pyproject.toml

[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "vai_lab"
authors = [
    {name = "Chris McGreavy", email = "chris.mcgreavy@aalto.fi"},
    {name = "Carlos Sevilla Salcedo", email = "carlos.sevillasalcedo@aalto.fi"}
]
readme = "README.md"
classifiers = [
    "License :: OSI Approved :: MIT License",
]
requires-python = ">=3.8"
version = "0.0.dev1"
description = "AI assisted Virtual Laboratory framework."

dependencies = [
    "numpy >= 1.20.0",
    "tk >= 0.1.0",
    "pillow >= 9.0.0", # package name of PIL
    "pandas >= 1.4.0",
    "sklearn >= 0.0",
    "ttkwidgets >= 0.12.0",
    "matplotlib >= 3.5.0"
]

[project.urls]
Source = "https://github.com/AaltoPML/VAI-labs"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"

[project.scripts]
aidesign = "aidesign.run_pipeline:main"